firebase_api_test/requirements.txt
CacheControl==0.12.5
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
Django==2.2.1
djangorestframework==3.9.4
drf-firebase-auth==0.0.4
firebase-admin==2.16.0
google-api-core==1.11.0
google-auth==1.6.3
google-cloud-core==1.0.0
google-cloud-firestore==1.2.0
google-cloud-storage==1.15.1
google-resumable-media==0.3.2
googleapis-common-protos==1.5.10
grpcio==1.20.1
idna==2.8
msgpack==0.6.1
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pytz==2019.1
requests==2.22.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.2
